[quant][pt2e][be] Separate xnnpack quantizer tests to separate file
Summary:
att

Test Plan:
python test/test_quantization.py TestQuantizePT2EXNNPACK
python test/test_quantization.py TestQuantizePT2EXNNPACKModels

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
